Decouple diffdisk (misnomer) from basedisk
Fix issue 1706.

Now basedisk is immediately renamed/converted to diffdisk
(not a diff despite the name) by the VM driver,
except when basedisk is an ISO9660 image.

Decoupling diffdisk from basedisk will eliminate the overhead of differencing I/O
and save some disk space.

I cannot remember why I designed the disk to be split into basedisk and diffdisk.
Maybe it was to allow `limactl factory-reset` to retain the initial state,
although the command was apparently never implemented in that way.
Maybe it was to allow multiple instances to share the same basedisk,
although it was never implemented in that way, either.

+++ pkg/driverutil/disk.go
@@ -19,6 +19,8 @@ import (
 )
 
 // EnsureDisk ensures that the diff disk exists with the specified size and format.
+// EnsureDisk usually just converts baseDisk (can be qcow2) to diffDisk (raw), unless baseDisk is an ISO9660 image.
+// Note that "diffDisk" is a misnomer, it is actually created as a full disk since Lima v2.1.
 func EnsureDisk(ctx context.Context, instDir, diskSize string, diskImageFormat image.Type) error {
 	diffDisk := filepath.Join(instDir, filenames.DiffDisk)
 	if _, err := os.Stat(diffDisk); err == nil || !errors.Is(err, os.ErrNotExist) {
@@ -29,33 +31,43 @@ func EnsureDisk(ctx context.Context, instDir, diskSize string, diskImageFormat i
 	diskUtil := proxyimgutil.NewDiskUtil(ctx)
 
 	baseDisk := filepath.Join(instDir, filenames.BaseDisk)
+	srcDisk := baseDisk
 
 	diskSizeInBytes, _ := units.RAMInBytes(diskSize)
 	if diskSizeInBytes == 0 {
 		return nil
 	}
-	isBaseDiskISO, err := iso9660util.IsISO9660(baseDisk)
-	if err != nil {
-		return err
-	}
-	srcDisk := baseDisk
-	if isBaseDiskISO {
-		srcDisk = diffDisk
-
-		// Create an empty data volume for the diff disk
-		diffDiskF, err := os.Create(diffDisk)
+	var isBaseDiskISO bool
+	if _, err := os.Stat(baseDisk); !errors.Is(err, os.ErrNotExist) {
+		isBaseDiskISO, err = iso9660util.IsISO9660(baseDisk)
 		if err != nil {
 			return err
 		}
+		if isBaseDiskISO {
+			srcDisk = diffDisk
 
-		if err = diffDiskF.Close(); err != nil {
+			// Create an empty data volume for the diff disk
+			diffDiskF, err := os.Create(diffDisk)
+			if err != nil {
+				return err
+			}
+
+			if err = diffDiskF.Close(); err != nil {
+				return err
+			}
+		}
+	}
+
+	// Check whether to use ASIF format
+
+	if err := diskUtil.Convert(ctx, diskImageFormat, srcDisk, diffDisk, &diskSizeInBytes, false); err != nil {
+		return fmt.Errorf("failed to convert %q to a disk %q: %w", srcDisk, diffDisk, err)
+	}
+
+	if !isBaseDiskISO {
+		if err := os.RemoveAll(baseDisk); err != nil {
 			return err
 		}
 	}
-	// Check whether to use ASIF format
-
-	if err = diskUtil.Convert(ctx, diskImageFormat, srcDisk, diffDisk, &diskSizeInBytes, false); err != nil {
-		return fmt.Errorf("failed to convert %q to a disk %q: %w", srcDisk, diffDisk, err)
-	}
-	return err
+	return nil
 }
